refactored Dashboard.js
-now all Search results belong to the SearchResults.js component
-now useLogin handles accessToken and refreshToken functionalities
