+++ src/Dashboard.js
@@ -1,20 +1,18 @@
 import axios from 'axios';
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import Footer from './components/Footer';
 import ParticlesLayer from './components/ParticlesLayer';
+import useLogin from './hooks/useLogin';
 import Player from './postLogin/Player';
 import SearchBar from './postLogin/SearchBar';
 import SearchResults from './postLogin/SearchResults';
-import SettingUpTimer from './postLogin/SettingUpTimer';
-import SongFrame from './postLogin/SongFrame';
 import Timer from './postLogin/Timer';
 
-
+//This component can be treated as the global storage of the app
 function Dashboard(props) {
     const code = props.code;
 
-  const [accessToken, setAccessToken] = useState(null);
-  const [refreshToken, setRefreshToken] = useState(null);
+  const accessToken = useLogin(code);
 
   //default green
   const [colour, setColour] = useState("#1DB954");
@@ -25,12 +23,13 @@ function Dashboard(props) {
   const [searchResult, setSearchResult] = useState(null);
   const [currentSongUri, setCurrentSongUri] = useState(null);
 
+  //when needed song is clicked
   const onSongPick = (songUri, albumColour) => {
-    setColour(albumColour)
+    setColour(albumColour);
     setCurrentSongUri(songUri);
   }
 
-  //search
+  //when changing word in the form. As we change, new set of songs appears
 const onChangeHandler = async (e) => {
   if (!e.target.value) return;
 
@@ -42,16 +41,7 @@ const onChangeHandler = async (e) => {
   setSearchResult(songs)
 }
 
-  //login
-  useEffect( () => {
-    axios.post('http://localhost:3001/login', {
-        code: code
-    }).then((result) => {
-      const bodyTokens = result.data.body;
-        setAccessToken(bodyTokens.access_token);
-  setRefreshToken(bodyTokens.refresh_token);
-    })
-    }, [])
+
 
   return (
     <>
@@ -76,32 +66,10 @@ const onChangeHandler = async (e) => {
 
     <br />
 
-    <div className='overflow-scroll w-1/2 my-2'>
-
-
-
-{searchResult && searchResult.map((song) => {
-  const image = song.album.images.reduce(
-    (smallest, image) => {
-      if (image.height > smallest.height) return image
-      return smallest
-    },
-    song.album.images[0]
-  )
-  return <SongFrame
-  key={song.uri}
-  uri={song.uri}
-  image={image.url}
-  onSongPick={onSongPick}
-  artist={song.artists[0].name}
-  title={song.name}
-  songTime={song.duration_ms} />
-})}
-
-</div>
-
-
-
+    <SearchResults
+      searchResult={searchResult}
+      onSongPick={onSongPick}
+    />
 
     </div>
     {/*4 seconds is the delay time*/}
